softap_transport: Send session number in HTTP cookies to track session Earlier, all request of SoftAP provisioning were not using same socket in some Android phones. This was indeed causing some failures for Wi-Fi provisioning.
To solve this issue, introduced http cookies on device side to track a session.
App will receive session number in cookie header and send it in next subsequent requests.
This change is backward compatible. If the device does not use cookies, android app will not send cookies.
This fix is available in provisioning library so updated just library version.