feature: Add command response feature for nodes
Implements a new feature to send commands to nodes and receive responses with status tracking.
This is just for testing out command-response workflow and can be found
under Node details, if the feature is enabled in firmware.
Production apps should integrate this elsewhere as per their requirement.

Note that this feature will be disabled by default and can be enabled
by setting isCommandResponseSupported=true in local.properties.